+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Changes
 
+- Kotlin 1.2.50
 - ViewModel binding dirty tracking for ListProperty (https://stackoverflow.com/questions/50364458/tornadofx-bind-dirty-properties-of-different-view-models)
 - Workspace.dock() will log a warning message if a child is docked while the workspace is not showing